After many tears, fixed conv input kernels.
It was another underflow bug in the kernel source code.

To aid in debugging, I'm adding extras/example_kernel.cc which just
contains some example code that gets auto-generated by this framework.
Normally this stage is transparent because the generated code gets
passed straight to the opencl compiler, however I'm adding it to show
what goes on behind the scenes.

+++ nnet/kernels/back_prop.kernel.cl
@@ -9,7 +9,7 @@
 // output: adjustment for the requested weight_index.
 double CalculateWeightGradient_LAYERID(global double* I, global double* W,
                                        const global double* GRADIENT,
-                                       size_t weight_index) {
+                                       int weight_index) {
   WEIGHT_GRADIENTS_HERE
 }
 
@@ -21,7 +21,7 @@ double CalculateWeightGradient_LAYERID(global double* I, global double* W,
 // output: gradient for the requested input_index.
 double CalculateInputGradient_LAYERID(global double* I, global double* W,
                                       const global double* GRADIENT,
-                                      size_t input_index) {
+                                      int input_index) {
   INPUT_GRADIENTS_HERE
 }
